docs/sample-apps/intelligence-hub/pyproject.toml
[project]
name = "intelligence-hub"
version = "0.1.0"
description = "Pixeltable Intelligence Hub -- a complete automation pipeline replacing n8n/Zapier"
requires-python = ">=3.10"
dependencies = [
    "pixeltable>=0.4",
    "openai>=1.0",
    "python-dotenv",
    "requests",
    "pandas",
    "sentence-transformers",
]

[project.optional-dependencies]
sheets = ["gspread"]
